move `config.go` to it's own package and make functions public. This way we have a single source of truth.
