+++ config/config.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package main
+package config
 
 import (
 	_ "embed"
@@ -23,13 +23,13 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-// getCertificatesDir return the directory where SSL certificates are saved
-func getCertificatesDir() *paths.Path {
-	return getDataDir()
+// GetCertificatesDir return the directory where SSL certificates are saved
+func GetCertificatesDir() *paths.Path {
+	return GetDataDir()
 }
 
-// getDataDir returns the full path to the default Arduino Create Agent data directory.
-func getDataDir() *paths.Path {
+// GetDataDir returns the full path to the default Arduino Create Agent data directory.
+func GetDataDir() *paths.Path {
 	userDir, err := os.UserHomeDir()
 	if err != nil {
 		log.Panicf("Could not get user dir: %s", err)
@@ -41,17 +41,17 @@ func getDataDir() *paths.Path {
 	return dataDir
 }
 
-// getLogsDir return the directory where logs are saved
-func getLogsDir() *paths.Path {
-	logsDir := getDataDir().Join("logs")
+// GetLogsDir return the directory where logs are saved
+func GetLogsDir() *paths.Path {
+	logsDir := GetDataDir().Join("logs")
 	if err := logsDir.MkdirAll(); err != nil {
 		log.Panicf("Can't create logs dir: %s", err)
 	}
 	return logsDir
 }
 
-// getDefaultConfigDir returns the full path to the default Arduino Create Agent configuration directory.
-func getDefaultConfigDir() *paths.Path {
+// GetDefaultConfigDir returns the full path to the default Arduino Create Agent configuration directory.
+func GetDefaultConfigDir() *paths.Path {
 	// UserConfigDir returns the default root directory to use
 	// for user-specific configuration data. Users should create
 	// their own application-specific subdirectory within this
@@ -79,13 +79,16 @@ func getDefaultConfigDir() *paths.Path {
 	return agentConfigDir
 }
 
+// https://github.com/golang/go/issues/46056
+//
+//go:generate cp -r ../config.ini config.ini
 //go:embed config.ini
 var configContent []byte
 
-// generateConfig function will take a directory path as an input
+// GenerateConfig function will take a directory path as an input
 // and will write the default config,ini file to that directory,
 // it will panic if something goes wrong
-func generateConfig(destDir *paths.Path) *paths.Path {
+func GenerateConfig(destDir *paths.Path) *paths.Path {
 	configPath := destDir.Join("config.ini")
 
 	// generate the config.ini file directly in destDir
