+++ systray/systray_real.go
@@ -21,12 +21,11 @@ package systray
 
 import (
 	"os"
-	"os/user"
 
 	log "github.com/sirupsen/logrus"
 
+	"github.com/arduino/arduino-create-agent/config"
 	"github.com/arduino/arduino-create-agent/icon"
-	"github.com/arduino/go-paths-helper"
 	"github.com/getlantern/systray"
 	"github.com/go-ini/ini"
 	"github.com/skratchdot/open-golang/open"
@@ -104,27 +103,19 @@ func (s *Systray) updateMenuItem(item *systray.MenuItem, disable bool) {
 
 // CrashesIsEmpty checks if the folder containing crash-reports is empty
 func (s *Systray) CrashesIsEmpty() bool {
-	logsDir := getLogsDir()
+	logsDir := config.GetLogsDir()
 	return logsDir.NotExist() // if the logs directory is empty we assume there are no crashreports
 }
 
 // RemoveCrashes removes the crash-reports from `logs` folder
 func (s *Systray) RemoveCrashes() {
-	logsDir := getLogsDir()
+	logsDir := config.GetLogsDir()
 	pathErr := logsDir.RemoveAll()
 	if pathErr != nil {
 		log.Errorf("Cannot remove crashreports: %s", pathErr)
 	} else {
 		log.Infof("Removed crashreports inside: %s", logsDir)
 	}
-}
-
-// getLogsDir simply returns the folder containing the logs
-func getLogsDir() *paths.Path {
-	usr, _ := user.Current()
-	usrDir := paths.New(usr.HomeDir) // The user folder, on linux/macos /home/<usr>/
-	agentDir := usrDir.Join(".arduino-create")
-	return agentDir.Join("logs")
 }
 
 // starthibernate creates a systray icon with menu options to resume/quit the agent
